Add complex rendering unit test for people list
It shouldn't really be there as it's not a true unit.
The table component handles rendering data and it should have
its own tests. In this app testing this kind of behaviour
should be done in e2e tests.

Still commiting as an exercise and for proper understanding
of component mounting in tests.